worker/requirements.txt
tornado==6.1
pynsq==0.9.0

# testing
pytest==6.1.2
ipython==7.19.0
ipdb==0.13.4
## The following requirements were added by pip freeze:
attrs==20.3.0
backcall==0.2.0
decorator==4.4.2
iniconfig==1.1.1
ipython-genutils==0.2.0
jedi==0.17.2
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
py==1.9.0
Pygments==2.7.2
pyparsing==2.4.7
six==1.15.0
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
